chore(refactor): extract useEditor into its own file
