+++ src/components/Editor/MarkdownContentEditor.vue
@@ -27,7 +27,8 @@ import { Editor } from '@tiptap/core'
 /* eslint-disable import/no-named-as-default */
 import History from '@tiptap/extension-history'
 import { getCurrentUser } from '@nextcloud/auth'
-import { provideEditor, ATTACHMENT_RESOLVER, editorFlagsKey } from '../Editor.provider.ts'
+import { ATTACHMENT_RESOLVER, editorFlagsKey } from '../Editor.provider.ts'
+import { provideEditor } from '../../composables/useEditor.ts'
 import { createMarkdownSerializer } from '../../extensions/Markdown.js'
 import AttachmentResolver from '../../services/AttachmentResolver.js'
 import markdownit from '../../markdownit/index.js'
