+++ src/components/SuggestionsBar.vue
@@ -61,7 +61,8 @@ import NcButton from '@nextcloud/vue/components/NcButton'
 import { Document, Shape, Upload, Table as TableIcon } from '../components/icons.js'
 import { useActionChooseLocalAttachmentMixin } from './Editor/MediaHandler.provider.js'
 import { getLinkWithPicker } from '@nextcloud/vue/dist/Components/NcRichText.js'
-import { useEditor, useFileMixin, useSyncServiceMixin } from './Editor.provider.ts'
+import { useFileMixin, useSyncServiceMixin } from './Editor.provider.ts'
+import { useEditor } from '../composables/useEditor.ts'
 import { generateUrl } from '@nextcloud/router'
 import { buildFilePicker } from '../helpers/filePicker.js'
 import { isMobileDevice } from '../helpers/isMobileDevice.js'
